refactor endpoint stats
