+++ app/docs/status_metadata.py
@@ -142,8 +142,7 @@ from various attacks and abuse patterns.</p>
     def get_metadata(cls) -> Dict[str, Any]:
         """Get complete metadata for FastAPI endpoint configuration."""
         return {
-            "tags": ["Security"],
-            "summary": "🛡️ Security System Status",
+            "summary": "Security System Status",
             "description": cls.get_full_description(),
             "response_description": "Current security system status and active protections",
             "responses": cls.get_responses()
